Various bug fixes - type now works.
One thing to note - the example template represents a stack which may fail deletion on the first pass due to the order in which the `InitialType` and `UpdatedType` are deleted. If this occurs, a second delete will succeed.

This version requires the following PRs from the Java Plugin to be published before this can be finalised;

* https://github.com/aws-cloudformation/cloudformation-cli-java-plugin/pull/287
* https://github.com/aws-cloudformation/cloudformation-cli-java-plugin/pull/288