Remove DocumentationUrl from readOnlyProperties
The ModuleVersion resource is currently not adhering to the [Resource type handler contract](https://docs.aws.amazon.com/cloudformation-cli/latest/userguide/resource-type-test-contract.html) specifically the rule that says that all readOnlyProperties must be returned in a read request. Theoretically, it would be a backwards-incompatible change to remove a property from readOnlyProperties, but since we have never returned this value in read requests in practice, this change won't break any existing customers.