[ModuleVersion] Fix intermittent contract test failures
There appears to be a bug that confuses the schema.json file, which is meant for a Module, with the resource schema. The contract tests end up trying to register a resource with a module schema which obviously fails. I tested this in our gamma environments and confirmed that the intermittent (but quite frequent) issue does not appear anymore.